interface/requirements.txt
customtkinter
Pillow
pandas
psycopg2-binary  # PostgreSQL
